+++ src/old/atomfeat.jl
